update path request in camera function
