--- III/amigos/amigos/soap/log.txt
@@ -1,1 +1,0 @@
-('PAN: {0}\n', 'TITL: {1}\n', 'ZOOM: 0.210938\n')
